+++ docs/changelog.md
@@ -26,6 +26,7 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 - Added the possibility to import and export individual menus via JSON files. Thanks to [@Yar2000T](https://github.com/Yar2000T) for this contribution!
 - Added the possibility to temporarily disable all Kando shortcuts while simulating hotkeys or executing macros. This is useful if you want to use a shortcut which is also used by Kando to trigger a menu. Surprisingly, this required a complete rewrite of how the binding of shortcuts works, so feel free to report any issues you encounter with shortcuts!
 - Added an option to disable achievement notifications. This can be found in the general settings dialog.
+- Added an option on Windows to disable the workaround for the issue where getting the stylus position is not possible with Windows Ink enabled. This workaround introduces a delay of 100ms before opening the menu, so if you don't use a stylus, you can disable it to make the menu open faster.
 - Some wlroots-specific options to handle the case where no pointer position is available. There are some timeouts which can now be configured in the settings dialog. Thanks to [@make-42](https://github.com/make-42) for this contribution!
 - Many translation updates: **Thanks to all the contributors!**